minor type safety improvements

+++ wiringPi/wiringPiI2C.h
@@ -22,6 +22,8 @@
  ***********************************************************************
  */
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -30,9 +32,9 @@ extern int wiringPiI2CRead           (int fd) ;
 extern int wiringPiI2CReadReg8       (int fd, int reg) ;
 extern int wiringPiI2CReadReg16      (int fd, int reg) ;
 
-extern int wiringPiI2CWrite          (int fd, int data) ;
-extern int wiringPiI2CWriteReg8      (int fd, int reg, int data) ;
-extern int wiringPiI2CWriteReg16     (int fd, int reg, int data) ;
+extern int wiringPiI2CWrite          (int fd, uint8_t data) ;
+extern int wiringPiI2CWriteReg8      (int fd, int reg, uint8_t data) ;
+extern int wiringPiI2CWriteReg16     (int fd, int reg, uint16_t data) ;
 
 extern int wiringPiI2CSetupInterface (const char *device, int devId) ;
 extern int wiringPiI2CSetup          (const int devId) ;
